Set up Devise and CanCan in dummy app
We need some kind of authentication framework for our blog.  After
several experiments, I've decided that we should use Devise and CanCan
in the _containing_ application, because that Just Works, and it gives
our users the option of single sign-on.

The other major alternative was to mount Devise at "/blog/admin", but
that got very nasty, very quickly.