Allow access to parent application helpers
I hit a tricky problem when intergrating rails_blog_engine into
http://kiddsoftware.com/.  I was trying to render the main application's
application.html.haml from inside a rails_blog_engine controller, and it
couldn't find the expected helpers.

To fix this, we need an overly-clever proxying system which detects when
the programmer tries to access an application helper from inside a
layout, and provides her with instructions on how to fix it.

This is too clever by half, but I really want to preserve the tight
integration between rails_blog_engine and the existing site's views.