Rename root_path/url to blog_path/url
We don't want to define root_path, because we may be rendering layouts
from our containing application, and if they contain a bare call to
'root_path', we want to catch this and ask the user to call
'app_helpers'.  But we can't do this if we shadow 'root_path'.